Move TopbarUserActions and its composable into src/layouts/
TopbarUserActions was only ever consumed by AppSimpleTopbar and
ManagerTopbar in src/layouts/components/, so it belongs there
rather than in the generic src/components/ root. Its composable
follows the same layout-scoped convention already used by
layout.ts / useMobileBarActiveState.ts / useRouteActiveMatch.ts.
