+++ src/layouts/components/TopbarUserActions.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import Button from 'primevue/button';
-import { useTopbarUserActions } from '@/composables/useTopbarUserActions';
+import { useTopbarUserActions } from '@/layouts/composables/useTopbarUserActions';
 
 const {
   t,
